Moving the generated code to the model package, as it makes more sense there.
