+++ dces-report-service/src/main/java/uk/gov/justice/laa/crime/dces/report/mapper/ContributionsFileMapper.java
@@ -5,10 +5,10 @@ import jakarta.xml.bind.JAXBException;
 import jakarta.xml.bind.Unmarshaller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import uk.gov.justice.laa.crime.dces.contributions.generated.ContributionFile;
-import uk.gov.justice.laa.crime.dces.contributions.generated.ContributionFile.CONTRIBUTIONSLIST.CONTRIBUTIONS;
-import uk.gov.justice.laa.crime.dces.contributions.generated.ContributionFile.CONTRIBUTIONSLIST.CONTRIBUTIONS.CcOutcomes.CcOutcome;
-import uk.gov.justice.laa.crime.dces.contributions.generated.ContributionFile.CONTRIBUTIONSLIST.CONTRIBUTIONS.Correspondence.Letter;
+import uk.gov.justice.laa.crime.dces.report.model.generated.ContributionFile;
+import uk.gov.justice.laa.crime.dces.report.model.generated.ContributionFile.CONTRIBUTIONSLIST.CONTRIBUTIONS;
+import uk.gov.justice.laa.crime.dces.report.model.generated.ContributionFile.CONTRIBUTIONSLIST.CONTRIBUTIONS.CcOutcomes.CcOutcome;
+import uk.gov.justice.laa.crime.dces.report.model.generated.ContributionFile.CONTRIBUTIONSLIST.CONTRIBUTIONS.Correspondence.Letter;
 import uk.gov.justice.laa.crime.dces.report.model.CSVDataLine;
 import uk.gov.justice.laa.crime.dces.report.service.CSVFileService;
 import uk.gov.justice.laa.crime.dces.utils.DateUtils;
