+++ dces-report-service/src/main/java/uk/gov/justice/laa/crime/dces/report/mapper/FdcFileMapper.java
@@ -6,7 +6,7 @@ import jakarta.xml.bind.JAXBException;
 import jakarta.xml.bind.Unmarshaller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import uk.gov.justice.laa.crime.dces.contributions.generated.FdcFile;
+import uk.gov.justice.laa.crime.dces.report.model.generated.FdcFile;
 import uk.gov.justice.laa.crime.dces.report.service.CSVFileService;
 
 import java.io.File;
